Better user path config and support for etworkd
add user path for configuration scan
check for NetworkManager/networkd
support for bullseye

(i.e. path: config/cli and files: packages, packages.remove)

+++ lib/distributions.sh
@@ -178,7 +178,11 @@ install_common()
 
 	else
 
-		cp "${SRC}/config/bootscripts/${bootscript_src}" "${SDCARD}/boot/${bootscript_dst}"
+		if [ -f "${USERPATCHES_PATH}/bootscripts/${bootscript_src}" ]; then
+			cp "${USERPATCHES_PATH}/bootscripts/${bootscript_src}" "${SDCARD}/boot/${bootscript_dst}"
+		else
+			cp "${SRC}/config/bootscripts/${bootscript_src}" "${SDCARD}/boot/${bootscript_dst}"
+		fi
 
 		if [[ -n $BOOTENV_FILE ]]; then
 			if [[ -f $USERPATCHES_PATH/bootenv/$BOOTENV_FILE ]]; then
@@ -450,7 +454,7 @@ install_common()
 	cp "${SDCARD}"/etc/armbian-release "${SDCARD}"/etc/armbian-image-release
 
 	# DNS fix. package resolvconf is not available everywhere
-	if [ -d /etc/resolvconf/resolv.conf.d ]; then
+	if [ -d /etc/resolvconf/resolv.conf.d ] && [ -n "$NAMESERVER" ]; then
 		echo "nameserver $NAMESERVER" > "${SDCARD}"/etc/resolvconf/resolv.conf.d/head
 	fi
 
@@ -460,31 +464,60 @@ install_common()
 	# enable PubkeyAuthentication
 	sed -i 's/#\?PubkeyAuthentication .*/PubkeyAuthentication yes/' "${SDCARD}"/etc/ssh/sshd_config
 
-	# configure network manager
-	sed "s/managed=\(.*\)/managed=true/g" -i "${SDCARD}"/etc/NetworkManager/NetworkManager.conf
+	if [ -f "${SDCARD}"/etc/NetworkManager/NetworkManager.conf ]; then
+		# configure network manager
+		sed "s/managed=\(.*\)/managed=true/g" -i "${SDCARD}"/etc/NetworkManager/NetworkManager.conf
 
-	# remove network manager defaults to handle eth by default
-	rm -f "${SDCARD}"/usr/lib/NetworkManager/conf.d/10-globally-managed-devices.conf
+		# remove network manager defaults to handle eth by default
+		rm -f "${SDCARD}"/usr/lib/NetworkManager/conf.d/10-globally-managed-devices.conf
 
-	# most likely we don't need to wait for nm to get online
-	chroot "${SDCARD}" /bin/bash -c "systemctl disable NetworkManager-wait-online.service" >> "${DEST}"/debug/install.log 2>&1
+		# most likely we don't need to wait for nm to get online
+		chroot "${SDCARD}" /bin/bash -c "systemctl disable NetworkManager-wait-online.service" >> "${DEST}"/debug/install.log 2>&1
+
+		# Just regular DNS and maintain /etc/resolv.conf as a file
+		sed "/dns/d" -i "${SDCARD}"/etc/NetworkManager/NetworkManager.conf
+		sed "s/\[main\]/\[main\]\ndns=default\nrc-manager=file/g" -i "${SDCARD}"/etc/NetworkManager/NetworkManager.conf
+		if [[ -n $NM_IGNORE_DEVICES ]]; then
+			mkdir -p "${SDCARD}"/etc/NetworkManager/conf.d/
+			cat <<-EOF > "${SDCARD}"/etc/NetworkManager/conf.d/10-ignore-interfaces.conf
+			[keyfile]
+			unmanaged-devices=$NM_IGNORE_DEVICES
+			EOF
+		fi
+
+	elif [ -d "${SDCARD}"/etc/systemd/network ]; then
+		# configure networkd
+		rm "${SDCARD}"/etc/resolv.conf
+		ln -s /run/systemd/resolve/resolv.conf "${SDCARD}"/etc/resolv.conf
+
+		# enable services
+		chroot "${SDCARD}" /bin/bash -c "systemctl enable systemd-networkd.service systemd-resolved.service" >> "${DEST}"/debug/install.log 2>&1
+
+		if  [ -e /etc/systemd/timesyncd.conf ]; then
+			chroot "${SDCARD}" /bin/bash -c "systemctl enable systemd-timesyncd.service" >> "${DEST}"/debug/install.log 2>&1
+		fi
+		umask 022
+		cat > "${SDCARD}"/etc/systemd/network/eth0.network <<- __EOF__
+		[Match]
+		Name=eth0
+
+		[Network]
+		DHCP=ipv4
+		LinkLocalAddressing=ipv4
+		#Address=192.168.1.100/24
+		#Gateway=192.168.1.1
+		#DNS=192.168.1.1
+		#Domains=example.com
+		NTP=0.pool.ntp.org 1.pool.ntp.org
+		__EOF__
+
+	fi
 
 	# avahi daemon defaults if exists
 	[[ -f "${SDCARD}"/usr/share/doc/avahi-daemon/examples/sftp-ssh.service ]] && \
 	cp "${SDCARD}"/usr/share/doc/avahi-daemon/examples/sftp-ssh.service "${SDCARD}"/etc/avahi/services/
 	[[ -f "${SDCARD}"/usr/share/doc/avahi-daemon/examples/ssh.service ]] && \
 	cp "${SDCARD}"/usr/share/doc/avahi-daemon/examples/ssh.service "${SDCARD}"/etc/avahi/services/
-
-	# Just regular DNS and maintain /etc/resolv.conf as a file
-	sed "/dns/d" -i "${SDCARD}"/etc/NetworkManager/NetworkManager.conf
-	sed "s/\[main\]/\[main\]\ndns=default\nrc-manager=file/g" -i "${SDCARD}"/etc/NetworkManager/NetworkManager.conf
-	if [[ -n $NM_IGNORE_DEVICES ]]; then
-		mkdir -p "${SDCARD}"/etc/NetworkManager/conf.d/
-		cat <<-EOF > "${SDCARD}"/etc/NetworkManager/conf.d/10-ignore-interfaces.conf
-		[keyfile]
-		unmanaged-devices=$NM_IGNORE_DEVICES
-		EOF
-	fi
 
 	# nsswitch settings for sane DNS behavior: remove resolve, assure libnss-myhostname support
 	sed "s/hosts\:.*/hosts:          files mymachines dns myhostname/g" -i "${SDCARD}"/etc/nsswitch.conf
@@ -584,15 +617,23 @@ install_distribution_specific()
 			# rc.local is not existing but one might need it
 			install_rclocal
 
-			# Basic Netplan config. Let NetworkManager manage all devices on this system
+			if [ -d "${SDCARD}"/etc/NetworkManager ]; then
+				local RENDERER=NetworkManager
+			else
+				local RENDERER=networkd
+			fi
+
+			# Basic Netplan config. Let NetworkManager/networkd manage all devices on this system
 			[[ -d "${SDCARD}"/etc/netplan ]] && cat <<-EOF > "${SDCARD}"/etc/netplan/armbian-default.yaml
 			network:
 			  version: 2
-			  renderer: NetworkManager
+			  renderer: $RENDERER
 			EOF
 
 			# DNS fix
-			sed -i "s/#DNS=.*/DNS=$NAMESERVER/g" "${SDCARD}"/etc/systemd/resolved.conf
+			if [ -n "$NAMESERVER" ]; then
+				sed -i "s/#DNS=.*/DNS=$NAMESERVER/g" "${SDCARD}"/etc/systemd/resolved.conf
+			fi
 
 			# Journal service adjustements
 			sed -i "s/#Storage=.*/Storage=volatile/g" "${SDCARD}"/etc/systemd/journald.conf
